[Improve][Connector] Improve hive connector with error data sink
