+++ datavines-connector/datavines-connector-plugins/datavines-connector-jdbc/src/main/java/io/datavines/connector/plugin/JdbcConfigBuilder.java
@@ -144,15 +144,7 @@ public class JdbcConfigBuilder implements ConfigBuilder {
 
     protected List<PluginParams> getOtherParams(boolean isEn) {
 
-        List<PluginParams> list = new ArrayList<>();
-
-        InputParam enableExternalCatalog = getInputParam("enable_use_view",
-                isEn ? "enable.use.view" : "允许使用视图",
-                isEn ? "please enter true or false" : "请填入 true 或者 false", 2, null,
-                "false");
-
-        list.add(enableExternalCatalog);
-        return list;
+        return null;
     }
 
 }
