+++ datavines-connector/datavines-connector-plugins/datavines-connector-jdbc/src/main/java/io/datavines/connector/plugin/utils/SqlUtils.java
@@ -14,19 +14,19 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.datavines.engine.local.api.utils;
+package io.datavines.connector.plugin.utils;
 
 import io.datavines.common.utils.StringUtils;
-import io.datavines.engine.local.api.entity.QueryColumn;
-import io.datavines.engine.local.api.entity.ResultList;
-import io.datavines.engine.local.api.entity.ResultListWithColumns;
+import io.datavines.connector.api.entity.QueryColumn;
+import io.datavines.connector.api.entity.ResultList;
+import io.datavines.connector.api.entity.ResultListWithColumns;
+import lombok.extern.slf4j.Slf4j;
 import net.sf.jsqlparser.JSQLParserException;
 import net.sf.jsqlparser.expression.Alias;
 import net.sf.jsqlparser.parser.CCJSqlParserUtil;
 import net.sf.jsqlparser.schema.Table;
 import net.sf.jsqlparser.statement.select.*;
 import org.apache.commons.collections4.CollectionUtils;
-import org.slf4j.Logger;
 
 import java.sql.*;
 import java.util.*;
@@ -34,9 +34,8 @@ import java.util.*;
 import static io.datavines.common.CommonConstants.DOT;
 import static org.apache.commons.lang3.StringUtils.EMPTY;
 
+@Slf4j
 public class SqlUtils {
-
-    protected static Logger log = LoggerFactory.getLogger(SqlUtils.class);
 
     public static ResultListWithColumns getListWithHeaderFromResultSet(ResultSet rs, Set<String> queryFromsAndJoins) throws SQLException {
