+++ datavines-engine/datavines-engine-plugins/datavines-engine-local/datavines-engine-local-transform/src/main/java/io/datavines/engine/local/transform/sql/ActualValueExecutor.java
@@ -18,8 +18,8 @@ package io.datavines.engine.local.transform.sql;
 
 import io.datavines.common.config.Config;
 import io.datavines.engine.local.api.LocalRuntimeEnvironment;
-import io.datavines.engine.local.api.entity.ResultList;
-import io.datavines.engine.local.api.utils.SqlUtils;
+import io.datavines.connector.api.entity.ResultList;
+import io.datavines.connector.plugin.utils.SqlUtils;
 import org.apache.commons.collections4.CollectionUtils;
 
 import java.sql.Connection;
